Add more tests and code clean up
Known issue: will put doc block one line too high when there is another
function above it. Not sure if this should be fixed, or if it should
be ran in conjunction with the fixer that fixes exactly that.

Also some general clean up, moved things to its own functions